monster speed
monsters speed should increase base on difficulty with what ive added to my test scene(ill have to test to verify)
