+++ Assets/Testing/JoshuaTesting/DifficultySelector.cs.meta
@@ -3,7 +3,10 @@ guid: 546e557c59bed834ca69db9c5ff071ef
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
-  defaultReferences: []
+  defaultReferences:
+  - difficultySetting: {instanceID: 0}
+  - monsterNavMeshAgent: {fileID: 8568769717320806197, guid: c217789bd399bde45ad9aa7b30e5d4e5,
+      type: 3}
   executionOrder: 0
   icon: {instanceID: 0}
   userData:
